solutions/enterprise-ai-agent/pyproject.toml
[project]
name = "enterprise-ai"
version = "0.1.0"
description = "The Enterprise AI agent."
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<4.0"
authors = [
    { name = "Pierrick Voulet" }
]
dependencies = [
    "google-adk (>=1.25.1,<2.0.0)",
    "google-api-core>=2.29.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.126.1",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "google-cloud-discoveryengine (>=0.13.12,<0.14.0)",
    "google-apps-chat (>=0.6.0,<0.7.0)",
]

[tool.poetry.group.dev.dependencies]
google-adk = { version = ">=1.18.0", extras = ["eval"] }

[tool.poetry.group.deployment.dependencies]
absl-py = ">=2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
